% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{.construct_api}
\alias{.construct_api}
\title{Standardized api object constructor}
\usage{
.construct_api(server, TFregulome_url, local_db_path)
}
\arguments{
\item{server}{server location from the calling function}

\item{TFregulome_url}{specified url from the calling function}

\item{local_db_path}{path to the local db from the calling function}
}
\value{
api_object
}
\description{
This function creates a standardized function to check if the TFregulome
server is valid. If that is successful it will construct a api-object and
return it. This is an internal helper function not meant to be called on
its own.
}
\examples{
.construct_api(server = "ca")
.construct_api(TFregulome_url = "https://methmotif.org/")
.construct_api(local_db_path = "/opt/database/tfregulome.sqlite")

}
\keyword{internal}
