% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMotif.R
\name{.searchMotif}
\alias{.searchMotif}
\title{Search motif PFM and beta score matrix (if source is MethMotif) for a given TFregulomeR ID in TFregulomeR}
\usage{
.searchMotif(id, motif_format = "MEME", api_object)
}
\arguments{
\item{id}{Required. TFregulomeR ID.}

\item{motif_format}{Motif PFM format, either in MEME by default or TRANSFAC.}

\item{api_object}{s4 object for interacting with the TFregulome database}
}
\value{
MethMotif class object
}
\description{
This function allows you to obtain motif PFM matrix and beta score matrix (if source is MethMotif) for a given TFregulomeR ID in TFregulomeR
}
\examples{
K562_CEBPB <- searchMotif(id = "MM1_HSA_K562_CEBPB")
K562_CEBPB_transfac <- searchMotif(id = "MM1_HSA_K562_CEBPB",
                                   motif_format = "TRANSFAC")
}
\keyword{MethMotif}
