% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPeaks.R
\name{loadPeaks}
\alias{loadPeaks}
\title{load peaks from TFregulomeR}
\usage{
loadPeaks(
  id,
  includeMotifOnly = FALSE,
  server = "ca",
  TFregulome_url,
  local_db_path = NULL
)
}
\arguments{
\item{id}{Required. TFregulomeR ID}

\item{includeMotifOnly}{Either TRUE or FALSE (default). If TRUE, only peaks with motif will be returned}

\item{server}{server localtion to be linked, either 'sg' or 'ca'.}

\item{TFregulome_url}{TFregulomeR server is implemented in MethMotif server. If the MethMotif url is NO more "https://bioinfo-csi.nus.edu.sg/methmotif/" or "https://methmotif.org", please use a new url.}

\item{local_db_path}{The complete path to the SQLite implementation of TFregulomeR database available at "https://methmotif.org/API_TFregulomeR/downloads/"}
}
\value{
a data.frame containing peak coordinates
}
\description{
This function allows you to obtain the peaks from TFregulomeR using TFregulomeR ID.
}
\examples{
CEBPB_peaks <- loadPeaks(id = "MM1_HSA_K562_CEBPB")
}
\keyword{loadPeaks}
