% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBrowser.R
\name{output_message}
\alias{output_message}
\title{A structured format for printing to the console}
\usage{
output_message(requested_df)
}
\arguments{
\item{requested_df}{data.frame containing the information of the queried TFBSs in TFregulomeR}
}
\value{
prints to console select metrics from the queried TFBSs in TFregulomeR
}
\description{
This function facilitates the printing of select metrics calculated from
the the queried TFBSs in TFregulomeR. This is an internal helper function
not meant to be called on its own.
}
\examples{
output_message(request_content_df_output)
}
\keyword{internal}
