% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTFBSTools.R
\name{toTFBSTools}
\alias{toTFBSTools}
\title{convert motif PFM in TFregulomeR into PFMatrix class object in TFBSTools package}
\usage{
toTFBSTools(id, server = "ca", TFregulome_url, local_db_path = NULL)
}
\arguments{
\item{id}{Required. TFregulomeR ID.}

\item{server}{server localtion to be linked, either 'sg' or 'ca'.}

\item{TFregulome_url}{TFregulomeR server is implemented in MethMotif server. If the MethMotif url is NO more "https://bioinfo-csi.nus.edu.sg/methmotif/" or "https://methmotif.org", please use a new url.}

\item{local_db_path}{The complete path to the SQLite implementation of TFregulomeR database available at "https://methmotif.org/API_ZIPPED.zip"}
}
\value{
An object of class PFMatrix
}
\description{
This function allows you to retrieve and convert motif PFM in TFregulomeR database into PFMatrix class object, which can be used in TFBSTools package.
}
\examples{
require(TFBSTools)
CEBPB_pfm <- toTFBSTools(id = "MM1_HSA_K562_CEBPB")
}
\keyword{toTFBSTools}
